Update BOM Header and WS2812B Rotation
When preparing a PCB order tonight, I needed two changes to the
existing script.  First, the header for the LCSC Part # field is now
expected to be "LCSC Part Number" per the instructions page[0].
Second, I've found that removing the rotation modifier yields the
correct orientation for the WS2812B LEDs that JLCPCB currently stocks
so I removed it from the database.

[0] - https://support.jlcpcb.com/article/84-how-to-generate-the-bom-and-centroid-file-from-kicad